+++ src/components/layermanager/LayerManagerPublisher.js
@@ -3,13 +3,18 @@ import PubSub from 'pubsub-js';
 import React from 'react';
 
 import {
-  STATUS_WARN, STATUS_INFO, IMAGES_ADD, MOLECULES_ADD, CELLS_ADD, CLEAR_PLEASE_WAIT,
+  STATUS_WARN, STATUS_INFO,
+  CELLS_ADD, CLUSTERS_ADD, FACTORS_ADD, GENES_ADD, IMAGES_ADD, MOLECULES_ADD, NEIGHBORHOODS_ADD,
+  CLEAR_PLEASE_WAIT,
 } from '../../events';
 
-import imagesSchema from '../../schemas/images.schema.json';
 import cellsSchema from '../../schemas/cells.schema.json';
+import clustersSchema from '../../schemas/clusters.schema.json';
+import factorsSchema from '../../schemas/factors.schema.json';
+import genesSchema from '../../schemas/genes.schema.json';
+import imagesSchema from '../../schemas/images.schema.json';
 import moleculesSchema from '../../schemas/molecules.schema.json';
-
+import neighborhoodsSchema from '../../schemas/neighborhoods.schema.json';
 
 function warn(message) {
   PubSub.publish(STATUS_WARN, message);
@@ -22,28 +27,47 @@ function info(fileName) {
 function loadLayer(layer) {
   const { name, type, url } = layer;
   const typeToSchema = {
-    IMAGES: imagesSchema,
     CELLS: cellsSchema,
+    CLUSTERS: clustersSchema,
+    FACTORS: factorsSchema,
+    GENES: genesSchema,
+    IMAGES: imagesSchema,
     MOLECULES: moleculesSchema,
+    NEIGHBORHOODS: neighborhoodsSchema,
   };
   const typeToEvent = {
-    IMAGES: IMAGES_ADD,
     CELLS: CELLS_ADD,
+    CLUSTERS: CLUSTERS_ADD,
+    FACTORS: FACTORS_ADD,
+    GENES: GENES_ADD,
+    IMAGES: IMAGES_ADD,
     MOLECULES: MOLECULES_ADD,
+    NEIGHBORHOODS: NEIGHBORHOODS_ADD,
   };
   fetch(url)
     .then((response) => {
       response.json().then((data) => {
-        const validate = new Ajv().compile(typeToSchema[type]);
+        const schema = typeToSchema[type];
+        if (!schema) {
+          throw Error(`No schema for ${type}`);
+        }
+        const validate = new Ajv().compile(schema);
         const valid = validate(data);
         if (valid) {
           PubSub.publish(typeToEvent[type], data);
           info(name);
         } else {
-          warn(`JSON from ${url} violates ${type} schema. Details in console.`);
-          console.warn(JSON.stringify(validate.errors, null, 2));
+          const failureReason = JSON.stringify(validate.errors, null, 2);
+          warn(`Error while validating ${name}. Details in console.`);
+          console.warn(`"${name}" (${type}) from ${url}: validation failed`, failureReason);
         }
+      }, (failureReason) => {
+        warn(`Error while parsing ${name}. Details in console.`);
+        console.warn(`"${name}" (${type}) from ${url}: parse failed`, failureReason);
       });
+    }, (failureReason) => {
+      warn(`Error while fetching ${name}. Details in console.`);
+      console.warn(`"${name}" (${type}) from ${url}: fetch failed`, failureReason);
     });
 }